+++ examples/at_modem/at_defs.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2020, Myriota Pty Ltd, All Rights Reserved
+// Copyright (c) 2020-2022, Myriota Pty Ltd, All Rights Reserved
 // SPDX-License-Identifier: BSD-3-Clause-Attribution
 //
 // This file is licensed under the BSD with attribution  (the "License"); you
@@ -50,6 +50,8 @@ enum {
   AT_CONTROL_RSSI,
   AT_CONTROL_SCHEDULE_MESSAGE,
   AT_CONTROL_SUSPEND_MODE,
+  AT_CONTROL_TIME,
+  AT_CONTROL_LOCATION,
   AT_CONTROL_NUM,
 };
 
@@ -68,8 +70,9 @@ enum { AT_STATE_INIT, AT_STATE_GNSS_ACQ, AT_STATE_READY, AT_STATE_UNKNOWN };
 static const char* Queries[] = {"MSGQ",    "STATE", "VSDK",     "MID",
                                 "REGCODE", "TIME",  "LOCATION", "SUSPEND"};
 
-static const char* Controls[] = {"SAVEMSG", "TXSTART", "TXSTOP", "GNSSFIX",
-                                 "RSSI",    "SMSG",    "SUSPEND"};
+static const char* Controls[] = {"SAVEMSG", "TXSTART", "TXSTOP",
+                                 "GNSSFIX", "RSSI",    "SMSG",
+                                 "SUSPEND", "TIME",    "LOCATION"};
 
 static const char* ErrorCodes[] = {
     "INVALID_PARAMETER", "MESSAGE_TOO_LONG",    "BUFFER_OVERFLOW",